requirements.txt
aiohttp==3.11.11
certifi==2024.12.14
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1       
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1        
eth-utils==5.1.0
eth_abi==5.1.0
loguru==0.7.3
openai==1.59.6
playwright==1.49.1
pyfiglet==1.0.2
pyuseragents==1.0.5
PyYAML==6.0.2
requests==2.32.3
web3==7.6.1